Add frame extractor for wildfire (#48)
* Add first version of frame extractor for wildfire

* comment version import, temporary solution for #49

* Add type annotations to cover FrameExtractor

* Fix frame extraction strategy not effectively passed to method

* FrameExtractor now only pick unique frames + Cover with tests + Add comments

Before, one may randomly select frame indexes with replacement so it resulted in a frame labeled multiple times in ~labels.csv but only one frame image found in folder. So it creates inconsistent count of frames registered and frames effectively available as images. Now the user get the count of frames he asked for, I think this is an expected behaviour. It raises exception when inconsistent number of frames is found. In the future, on may relax this by maybe skipping the state that has not enough frames to cover the user needs, but still provide expected number of frames per state

* Add Integration tests for FrameExtractor

* Test for multiple frames

* Raise Exception when Video frame unreadable + cover with test

* Print in Exception from which video they is inconsistent number of frames

* Allow frames duplicates when too many asked to be extracted (not default behaviour)

it complies with previous version of the extractor, so it now raises warnings when allowed

* Add automatic download of video fixtures

* Add new requirements

* add yaml dependency

* Convert type for states start and End

Because you are using "range" and not "arange" at 
# Trying to set a valid frame range
frames_range = range(state.stateStart, state.stateEnd + 1)

you need stateStart and stateEnd to be int and not float as it is initially in the states dataframe

* correct typing error in my last commit

self and not this

* Add the posibility to resize the frame

I added an option to resize the frame before writing in order to speed up the code. You can use it giving:

resizeFrame=(width, height) to the run method

* Fix pkg-style

* lower bound packages version to ensure backward compatibility

* Strictly comply with PEP8 concerning imports permitting download that may break in the future

* Revert reviewer commits in favor of another PR

Co-authored-by: lostanlen <mateo.lostanlen@gmail.com>